% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepReI.R
\name{DeepReI}
\alias{DeepReI}
\title{Retention index predictions}
\usage{
DeepReI()
}
\arguments{
\item{...}{function has no parameters (data input is performed by GUI)}
}
\value{
shiny.appobj
}
\description{
\code{DeepReI} (Deep Learning-based Retention Index predictor) is gas chromatographic retention index predictor based on convolutional neural network.
}
\details{
It is designed to predict retention index of molecules (represented as SMILES strings) on SEMI-STANDARD NON-POLAR stationary
phases (e.g. 5% phenyl(dimethyl)polysiloxane). Data input, prediction and export of results is performed through GUI of shiny based app.
}
\references{
Tomáš Vrzal, Michaela Malečková, Jana Olšovská,DeepReI: Deep learning-based gas chromatographic retention index predictor,Analytica Chimica Acta 1147, 2021, 64 - 71
}
\author{
Tomas Vrzal <tomas.vrzal@beerresearch.cz>
}
